Redirected the Assistant to stderr, ensuring that it and the traceback are printed within the same buffer.
Changed the sizing method, taking the best of manual formatting and automatic formatting. `traceback` doesn't use cpythons current format for determining the pathname of an object, but it does format most of the messages correct.

Small limitation with 3.10 in mind, the suggestions added by 3.10's error handling update doesn't cross over to python's traceback. Namely the suggestion system added is done completely in python, negating the ability to grab it. The result of STDERR could be captured to parse that first, ensuring that it's always the correct length.